+++ BUILD.gn
@@ -1,5 +1,5 @@
 # Copyright (C) 2018-2019 The ANGLE Project Authors.
-# Copyright (C) 2019 LunarG, Inc.
+# Copyright (C) 2019, 2023 LunarG, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -64,8 +64,14 @@ if (!is_android) {
     public_deps = [ "$vulkan_headers_dir:vulkan_headers" ]
     data_deps = [ ":vulkan_gen_icd_json_file" ]
     sources = [
-      "icd/generated/mock_icd.cpp",
-      "icd/generated/mock_icd.h",
+      "icd/mock_icd.h",
+      "icd/mock_icd.cpp",
+      "icd/generated/function_declarations.h",
+      "icd/generated/function_definitions.h",
+    ]
+    include_dirs = [
+        "icd",
+        "icd/generated",
     ]
     if (is_win) {
       sources += [ "icd/VkICD_mock_icd.def" ]
