+++ scripts/generate_source.py
@@ -46,8 +46,8 @@ def main(argv):
 
     # output paths and the list of files in the path
     files_to_gen = {str(os.path.join('icd','generated')) : ['vk_typemap_helper.h',
-                                            'mock_icd.h',
-                                            'mock_icd.cpp'],
+                                            'function_definitions.h',
+                                            'function_declarations.h'],
                     str(os.path.join('vulkaninfo','generated')): ['vulkaninfo.hpp']}
 
     #base directory for the source repository
